requirements.txt
streamlit
google-genai
python-dotenv
Pillow
gTTS
